one-off script for setting video metadata (content-disposition)
You need to set the content-disposition:attachment metadata to force the
object to download.  This little boto one-off accomplishes that.

Would be much nicer if this got the prior metadata and then only changed
the value that we wanted to change, but I don't need that right now, and
I'd rather sleep.