Don't double print the error message in the title.
Pytest files don't have a 'type' element but do have a 'message'
element. So the logic here would first extract the error message from
the error output and then append to it the name that it gets from the
message attribute.  These two items were equal so the error line was the
same message repeated twice with at colon between them.